pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ddae-seismic"
version = "0.1.0"
description = "Refactored DDAE pipeline for seismic random noise attenuation"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "matplotlib>=3.7",
  "pyyaml>=6.0",
  "tensorflow>=2.15",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
